[model] moved SShapeElement to bounds feature and removed unnnecessary superclasses [button] extracted button feature
