+++ client/examples/classdiagram/src/di.config.ts
@@ -1,3 +1,4 @@
+import { } from '../../../lib/src/features/button/di.config';
 /*
  * Copyright (C) 2017 TypeFox and others.
  *
@@ -10,7 +11,7 @@ import {
     defaultModule, TYPES, ViewRegistry, overrideViewerOptions, SGraphView, SLabelView, SCompartmentView,
     PolylineEdgeView, ConsoleLogger, LogLevel, WebSocketDiagramServer, boundsModule, moveModule, selectModule,
     undoRedoModule, viewportModule, hoverModule, LocalModelSource, HtmlRootView, PreRenderedView, 
-    exportModule, expandModule, fadeModule, ExpandButtonView
+    exportModule, expandModule, fadeModule, ExpandButtonView, buttonModule
 } from "../../../src"
 import { ClassNodeView, IconView} from "./views"
 import { ClassDiagramFactory } from "./model-factory"
@@ -27,7 +28,9 @@ const classDiagramModule = new ContainerModule((bind, unbind, isBound, rebind) =
 
 export default (useWebsocket: boolean, containerId: string) => {
     const container = new Container()
-    container.load(defaultModule, selectModule, moveModule, boundsModule, undoRedoModule, viewportModule, fadeModule, hoverModule, exportModule, expandModule, classDiagramModule)
+    container.load(defaultModule, selectModule, moveModule, boundsModule, 
+        undoRedoModule, viewportModule, fadeModule, hoverModule, 
+        exportModule, expandModule, buttonModule, classDiagramModule)
     if (useWebsocket)
         container.bind(TYPES.ModelSource).to(WebSocketDiagramServer).inSingletonScope()
     else
