+++ client/src/features/button/di.config.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright (C) 2017 TypeFox and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+import { ContainerModule } from "inversify"
+import { ButtonHandlerRegistry } from './button-handler'
+
+const buttonModule = new ContainerModule(bind => {
+    bind(ButtonHandlerRegistry).toSelf().inSingletonScope()
+})
+
+export default buttonModule
