+++ client/src/features/select/di.config.ts
@@ -8,13 +8,11 @@
 import { ContainerModule } from "inversify"
 import { TYPES } from "../../base/types"
 import { SelectCommand, SelectKeyboardListener, SelectMouseListener } from "./select"
-import { ButtonHandlerRegistry } from './button-handler'
 
 const selectModule = new ContainerModule(bind => {
     bind(TYPES.ICommand).toConstructor(SelectCommand)
     bind(TYPES.KeyListener).to(SelectKeyboardListener)
     bind(TYPES.MouseListener).to(SelectMouseListener)
-    bind(ButtonHandlerRegistry).toSelf().inSingletonScope()
 })
 
 export default selectModule
