+++ client/src/features/select/select.ts
@@ -16,9 +16,9 @@ import { MouseListener } from "../../base/views/mouse-tool"
 import { KeyListener } from "../../base/views/key-tool"
 import { setClass } from "../../base/views/vnode-utils"
 import { isSelectable } from "./model"
-import { ButtonHandlerRegistry } from './button-handler'
-import { inject } from 'inversify'
-import { SButton } from '../../graph/sgraph'
+import { ButtonHandlerRegistry } from '../button/button-handler'
+import { inject, optional } from 'inversify'
+import { SButton } from '../button/model'
 
 /**
  * Triggered when the user changes the selection, e.g. by clicking on a selectable element. The resulting
@@ -165,7 +165,7 @@ export class SelectCommand extends Command {
 
 export class SelectMouseListener extends MouseListener {
 
-    constructor(@inject(ButtonHandlerRegistry) protected buttonHandlerRegistry: ButtonHandlerRegistry) {
+    constructor(@inject(ButtonHandlerRegistry)@optional() protected buttonHandlerRegistry: ButtonHandlerRegistry) {
         super()
     }
 
@@ -174,7 +174,7 @@ export class SelectMouseListener extends MouseListener {
 
     mouseDown(target: SModelElement, event: MouseEvent): Action[] {
         if (event.button === 0) {
-            if (target instanceof SButton && target.enabled) {
+            if (this.buttonHandlerRegistry !== undefined && target instanceof SButton && target.enabled) {
                 const buttonHandler = this.buttonHandlerRegistry.get(target.type)
                 if (buttonHandler !== undefined)
                     return buttonHandler.buttonPressed(target)
